+++ src/Infrastructure/Persistence/Configurations/AppUserConfiguration.cs
@@ -1,0 +1,13 @@
+using Core.Domain.Entities;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore.Metadata.Builders;
+
+namespace Infrastructure.Persistence.Configurations;
+
+public class AppUserConfiguration : IEntityTypeConfiguration<AppUser>
+{
+    public void Configure(EntityTypeBuilder<AppUser> builder)
+    {
+        builder.HasIndex(u => u.CreatedAt).HasDatabaseName("ix_users_createdat");
+    }
+}